+++ lib/xwidget.dart
@@ -1,17 +1,18 @@
-export 'src/custom/async.dart';
+export 'package:xwidget_el/xwidget_el.dart';
+
 export 'src/analytics/analytics.dart';
+export 'src/custom/async.dart';
 export 'src/custom/collection.dart';
 export 'src/custom/controller.dart';
 export 'src/custom/event_listener.dart';
 export 'src/custom/material_state.dart';
 export 'src/custom/media.dart';
 export 'src/custom/value_listener.dart';
+export 'src/defaults.dart';
 export 'src/utils/events.dart';
 export 'src/utils/parsers_material.dart';
 export 'src/utils/resources/resources.dart';
-export 'src/utils/utils.dart';
-export 'src/defaults.dart';
+export 'src/utils/router/web_url_sync.dart';
 export 'src/utils/router/xrouter.dart';
+export 'src/utils/utils.dart';
 export 'src/xwidget.dart';
-
-export 'package:xwidget_el/xwidget_el.dart';